+++ src/features/Interview/sampleData.ts
@@ -1,0 +1,50 @@
+export const dataSample = [
+    {
+      NIC: "200112345678",
+      name: "John Doe",
+      date: "2025-05-01",
+      duration: "3 weeks",
+      departmentId: [1, 3],
+      department: ["Software Engineering", "AI Research"]
+    },
+    {
+      NIC: "200212345679",
+      name: "Jane Smith",
+      date: "2025-05-03",
+      duration: "1 month",
+      departmentId: [2],
+      department: ["Data Science"]
+    },
+    {
+      NIC: "199812345678",
+      name: "Michael Lee",
+      date: "2025-05-05",
+      duration: "2 weeks",
+      departmentId: [4, 5],
+      department: ["Cybersecurity", "Networking"]
+    },
+    {
+      NIC: "200012345689",
+      name: "Amara Silva",
+      date: "2025-05-07",
+      duration: "6 months",
+      departmentId: [3],
+      department: ["AI Research"]
+    },
+    {
+      NIC: "199912345610",
+      name: "Ravi Perera",
+      date: "2025-03-10",
+      duration: "1 year",
+      departmentId: [1, 2, 4],
+      department: ["Software Engineering", "Data Science", "Cybersecurity"]
+    },
+    {
+        NIC:"973070070V",
+        name:"John cena",
+        date:"2025-05-01",
+        duration:"3 weeks",
+        departmentId:[1, 3],
+        department:["Software Engineering", "AI Research"]
+    }
+];
